requirements.txt
lxml
requests
discord.py
python-decouple
beautifulsoup4
discord-components
discord-interactions
